use new pipeline config, impl Write trait

+++ src/uart0/interface.rs
@@ -2,24 +2,25 @@
  * Copyright (c) 2019 by the authors
  *
  * Author: André Borrmann
- * License: Apache License 2.0
+ * License: MIT / Apache License 2.0
  **********************************************************************************************************************/
 
 //! # Low-Level Uart0 interface implementation
 //!
 
 use ruspiro_gpio::GPIO;
-use ruspiro_register::{define_mmio_register, RegisterFieldValue};
+use ruspiro_mmio_register::define_mmio_register;
+use ruspiro_register::RegisterFieldValue;
 use ruspiro_timer as timer;
 
 use crate::UartResult;
 
 // Peripheral MMIO base address - depends on the right feature
 #[cfg(feature = "ruspiro_pi3")]
-const PERIPHERAL_BASE: u32 = 0x3F00_0000;
+const PERIPHERAL_BASE: usize = 0x3F00_0000;
 
 // UART0 MMIO base address
-const UART0_BASE: u32 = PERIPHERAL_BASE + 0x0020_1000;
+const UART0_BASE: usize = PERIPHERAL_BASE + 0x0020_1000;
 
 /// Initialize the Uart0 based on the given core rate and baud rate.
 /// For the time beeing the Uart0 will be bridged to the Raspberry Pi
@@ -158,7 +159,7 @@ define_mmio_register![
         INT_RX      OFFSET(4), // receive FiFo reached water mark
         INT_DSRM    OFFSET(3),
         INT_DCDM    OFFSET(2),
-        INT_CTSM    OFFSET(1)     
+        INT_CTSM    OFFSET(1)
     },
     UART0_RIS<ReadWrite<u32>@(UART0_BASE + 0x3C)>,
     UART0_MIS<ReadWrite<u32>@(UART0_BASE + 0x40)>,
